==> member-ricky/src/2_calculate_score_table.ipynb <==
# Generate TER Table from specified dataset
import pandas as pd

df = pd.read_csv("../../data/preprocessed_data/subsampled_preprocessed.csv")

df.head()

# Convert string type to datetime
df['tpep_pickup_datetime'] = pd.to_datetime(df['tpep_pickup_datetime'])
df['tpep_dropoff_datetime'] = pd.to_datetime(df['tpep_dropoff_datetime'])

# Total earninigs = tip_amount + total_amount
df['total_earnings'] = df["tip_amount"] + df['total_amount']

# Calculate trip_duration in seconds
df["trip_duration"] = (df['tpep_dropoff_datetime'] - df['tpep_pickup_datetime']).dt.total_seconds()

# Extract day of week and minute of day
df["day_of_week"] = df["tpep_pickup_datetime"].dt.weekday.astype(str)
df["minute_of_day"] = df["tpep_pickup_datetime"].dt.time

# Round down time to nearest minute
df['minute_of_day'] =  [i.replace(second=0) for i in df['minute_of_day']]

# Compute frequency and TER
df_ter = df.groupby(["PickupCell", "day_of_week", "minute_of_day"]) \
            .size() \
            .to_frame(name='frequency') \
            .join(df.groupby(["PickupCell", "day_of_week", "minute_of_day"]) \
                    .agg({"total_earnings":"mean", "trip_duration":"mean"}) \
                    .rename(columns={"trip_duration":"avg(td)", "total_earnings":"avg(te)"}) \
            ).reset_index()

df_ter['avg(ter)'] = df_ter['avg(te)']/df_ter['avg(td)']

df_ter.to_csv("../csv/df_score_table.csv", index=False)

df_ter['day_of_week'][0]
