==> member-ricky/src/3_sklearn_glm.ipynb <==
%reset

import pandas as pd
import pickle

# import csv, specific some columns as strings
df = pd.read_csv("../../data/preprocessed_data/subsampled_preprocessed.csv")
#                  dtype={'Pickup_day':object,
#                         'day_of_week': object,
#                        'Pickup_hour':object,
#                        'Pickup_minute':object,
#                        'Pickup_week':object})

attrs1 = ['Season', 'Pickup_month', 'Pickup_week', 'Pickup_day', 'Pickup_wday',
       'Weekend', 'Pickup_hour', 'Time', 'Pickup_minute', 'Zone',
       'Pickup_cell', 'Number_trips']

attrs2 = ['Pickup_week', 'Pickup_wday', 'Pickup_hour', 'Pickup_minute','Pickup_cell', 'Number_trips']

attrs3 = ['Pickup_cell', 'Number_trips']

df2['Pickup_minute'] = df2['Pickup_minute'].apply(lambda x: x//5 * 5)

df2 = df2.groupby(attrs1[:-1]).sum().reset_index()

### Get X and y

from sklearn.preprocessing import OneHotEncoder

attrs = attrs1

df2 = df.copy()[attrs]
_X = df2[attrs[0:len(attrs)-1]]
y = df2["Number_trips"].astype('float')

# One hot encode X,return sparse matrix
ohe = OneHotEncoder(handle_unknown='ignore')
ohe.fit(_X)

X = ohe.transform(_X)

with open('../players/model/ohe.pkl', 'wb') as handle:
    pickle.dump(ohe, handle, protocol=pickle.HIGHEST_PROTOCOL)

def get_reg_train(attrs):
    
    df2 = df.copy()[attrs]
    X = df2[attrs[0:len(attrs)-1]]
    y = df2["Number_trips"].astype('float')

    # One hot encode X,return sparse matrix
    ohe = OneHotEncoder()
    X = ohe.fit_transform(X)
    
    return (X, y)

## Model training with cross validation

from sklearn.model_selection import train_test_split

X, y = get_reg_train(attrs)
X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.1, random_state=42)

### Multinomial Naive Bayes

from sklearn.naive_bayes import MultinomialNB

# mnb = MultinomialNB()
# mnb.fit(X_train, y_train)

mnb = MultinomialNB()
mnb.fit(X, y)

mnb.score(X_test, y_test)

with open('../players/model/mnb.pkl', 'wb') as handle:
    pickle.dump(mnb, handle, protocol=pickle.HIGHEST_PROTOCOL)

mnb.predict(test)

### cross validation score
* 1: 0.5569597379648528
* 2: 0.5641991817107989
* 3: 0.5640146767072894

### Linear Regression

from sklearn import linear_model
from sklearn.metrics import mean_squared_error

for i in [attrs1, attrs2, attrs3]:
    lr = linear_model.LinearRegression()
    X, y = get_reg_train(i)
    lar.fit(X, y)
    y_pred = lar.predict(X)
    print(i,"=",mean_squared_error(y_pred, y))

y_pred = lr.predict(X)
mean_squared_error(y_pred, y)

### mean square error

1. 1.482783895638988
2. 1.8353711646395263
3. 1.5895257371882623

### Lasso Regression

from sklearn import linear_model
from sklearn.metrics import mean_squared_error

for i in [attrs1, attrs2, attrs3]:
    lar = linear_model.Lasso(alpha=0.01)
    X, y = get_reg_train(i)
    lar.fit(X, y)
    y_pred = lar.predict(X)
    print(i,"=",mean_squared_error(y_pred, y))

for i in [attrs1, attrs2, attrs3]:
    lar = linear_model.Lasso(alpha=0.1)
    X, y = get_reg_train(i)
    lar.fit(X, y)
    y_pred = lar.predict(X)
    print(i,"=",mean_squared_error(y_pred, y))

### Poisson regression

# Poisson regression
import statsmodels.api as sm
import statsmodels.formula.api as smf

formula = 'Number_trips~ Pickup_week+Pickup_wday+Pickup_hour+Pickup_minute+Pickup_cell

data = df2

pois_glm = smf.glm(formula=formula, data=data, family=sm.families.Poisson()).fit()
